backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pillow==10.1.0
python-dotenv==1.0.0